Pt_2: Error correction
Pt_2_1:
Fixed errors with the number of characters on the lines.

Pt_2_2:
Fixed errors with the number of characters on the lines.

Pt_2_3:
Fixed errors with the number of characters on the lines.

Pt_2_4:
Fixed errors with the number of characters on the lines.
Fixed errors with spaces between characters.

Pt_2_5:
Fixed errors with the number of characters on the lines.

Pt_2_6:
Fixed errors with the number of characters on the lines.
Fixed errors with spaces between characters.

Pt_2_7:
Fixed errors with the number of characters on the lines.
W292 and W293 (?)

Pt_2_8:
Fixed errors with the number of characters on the lines.

Pt_2_9:
Fixed errors with spaces between characters.

Pt_2_10:
Fixed errors with spaces between characters.

Pt_2_11:
Fixed errors with spaces between characters.